+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 
 * Puppeteer updated to v3
 * DragAndDrop method
-* Other dependencies updated
+* Minor fix in types
+* Dependencies updated
 
 2.11.3 / 2020-03-07
 ===================